refactored validation schema to work as expected in server
