+++ app/api/auth/verify-user/route.js
@@ -1,5 +1,5 @@
-import { authValidationSchema } from '@/components/auth/auth-form';
 import prisma from '@/lib/db/prisma';
+import { authValidationSchema } from '@/lib/schema/auth-schema';
 import { NextResponse } from 'next/server';
 
 export async function POST(req) {
